Vampires get their regenerate back and transfix fix for thralls. (#2105)
## About The Pull Request
This is my first ever PR ever, I don't know if it will be considered or
not but it was fun fucking around with code and github.

This PR does 3 things:
1- readds regenerate for vampires
2- gives converted vampire spawns average blood sorcery so they can use
their transfix (they need blood sorcery for it to work)
3- gives vampires the attach limb trait so they can reattach their
severed arms or legs for aura farming purposes.

## Why It's Good For The Game
I feel like vampires are in a pretty tough spot compared to liches and
werewolves and the silver weakness of the vampire now works properly.
Regeneration is in my opinion a vital part of a vampire, playing a
vampire spawn currently feels nearly identical to playing a skeleton.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->