Zizocult Skill Changes (#1999)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Originally, every Zizoid lackey gained a +2 to sword and knife skills,
this is no more, cry harder, pick your converts well so you dont have to
rely on flat bonuses to skill. The Zizoid cult leader would get a
whopping +4 to knives, swords, athletics, and a +5 to wrestling, leading
to you having legendary in all of those skills if you were a class that
had decent of those skills before, now it just sets your skill in those
to 4, 4, 5, 4 respectively.

## Why It's Good For The Game

Prevents lackeys and the leader from being given legendary skills, the
possibility of having a group of like 4 legendary/master fighters is
bad, this also makes it so drow knight no longer gets legendary sword
skill as they were also affected by this.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->