Youngling Nerf (#2094)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Give a huge modifier to Youngling about carrying armors. Along with
that, They are now unable to get the halve str requirement from wielding
a weapon.

![image](https://github.com/user-attachments/assets/57b8b983-6cf2-4b9a-a7fc-92e64987d793)

![image](https://github.com/user-attachments/assets/dc752ef4-01a2-48af-bbaa-d733e2ddec38)

![image](https://github.com/user-attachments/assets/97f1522e-26f4-4c75-8883-d812c3254f28)

Harpies now get a innate +1 in music, and their vocal cords allows them
to play songs and SHOULD give them +1 in music too.
Harpies also get their endurance malus removed for upcoming changes to
flight.
Harpies get their critical weakness removed.

![image](https://github.com/user-attachments/assets/a863dea9-bf2f-4475-a3c3-b9e6f246154f)

![image](https://github.com/user-attachments/assets/94e028f7-aab0-4d2b-9081-054d692726fe)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
No comments about the youngling.
Aberra wished for harpies to get those things. I plan to also change
flight to also drain stamina, but i'm genuinely unsure how i'll go with
this.
Huge thanks to cheffie and grung for helping me fix this
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->