Masochist update and rebalance + redundant flaw removal (#2089)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->
Removes pain freek entirely for its redundancy.

Reduced the pain requirement for masochist to be a lot less extreme.

Added an endurance-based modifier to the amount of pain required,
capping at 50% for endurance 5 and 150% for endurance 15.


## Why It's Good For The Game
Pain freek and masochist co-existing was confusing and unintuitive, with
one of them acting as scream activated and the other caused by specific
levels of pain, only one of them needed to exist.

Masochist had two major problems with its pain calculation. For
starters, it didn't take into consideration the pain
tolerance(endurance) of the character. This means that a tiny frail
wizard and a monk had to reach the same level of pain to get the same
rewards, given how pain tolerance effects how you experience pain, this
made little sense. Additionally, the amount of pain required to satisfy
the flaw was frankly absurd. Every 25 or less minutes you would have to
_almost_ kill yourself just to reach a high enough pain level. Once you
had reached this level you were most likely constantly flopping over
from pain (unless you used blood loss primarily, but good luck doing
that safely.)


<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->