Vagabond Hair (The Re-Vagaing) (#2150)
![image](https://github.com/user-attachments/assets/ca6bcc46-ee5c-43f6-a125-68147eb05a01)

<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Makes alignment adjustments and sprite edit to the vagabond hairstyle
for roundstart
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Quality of life and adjustment to make it look better
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ Checked] You tested this on a local server.
- [Checked ] This code did not runtime during testing.
- [Checked ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->