Allows the Inhumen Pantheon to be Clerics / Adds the respective gods as Templar/Cleric picks / Gives Xylix Templar a song (#2139)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

CHANGES:

Templar (Xylix) now has a song assigned to him. I saw this issue working
on the PR and plugged it in.

The 4 are now added to the list of possible clerics.

Clerics & Paladins now have heretical god variants, complete with being
branded as a heretic.

Clerics & Paladins now have their respective god's music. They also get
-1 and -2 LCK respectively.

_For Paladins - Baothans get a crown and sunglasses. Graggarites get the
Sinistar. Matthios worshippers get the Hedge Knight helmet, Zizo
worshippers get the Lich hood.

All heretics do not have any religious iconography, they won't be able
to cast spells off the bat. This is not an oversight. Suffer dipshit.

## Why It's Good For The Game

Previously if you played paladin as a heretic you got -20 luck. It's
most certainly antithetical to play such a role as one, but I think
freedom of choice overrules this. If people want to play a heretic, fuck
it, let em. Gives the forest garrison more work to do anyways.

Paladin is capped at 1, Cleric is capped at 4. I doubt this will create
issues with people using these roles as an excuse to be bandit-lite.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->