(SOUND) Removes all deprecated/unused music (#2197)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Ditto. Read title

## Why It's Good For The Game

These songs were never meant to be utilized for anything. Early on in
the codebase we made the decision to move away from copywritten music /
tracks we never sought permission for. They were kept in the code as
it'd be nice for folks looking for the previously used songs to find em,
plus I think seeing how things once were is just neat overall

A few people have made PR's to use these tracks, they're not meant to be
used. Unused was the wrong term for it, and looking out for folks using
em is a little tedious. Sorry, killing em. Goodbye desert_rider_combat,
you were great

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->