Assassin - Remove Whitecloak Disguise (#2196)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Removes the Whitecloak assassin disguise. Requested change

## Why It's Good For The Game

There's another PR for a version of this that has far more effort - [The
Faceless.](https://github.com/Monkestation/Vanderlin/pull/2021) There's
also some issues with escalation/moderation, and people dropping the
disguise for stealth - surpassing the entire point of the outfit in the
first place.

Perhaps someday the Whitecloak will return, but for now, it's being put
on a raft and kicked out to sea. Go aura-farm in the great barrier reef
dipshit

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->