Language Icons Show If Even If Understood (#2191)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Makes the language icons that show when you don't understand a language,
show even if you do speak that language. Does this for all standard
languages except imperial (common).

## Why It's Good For The Game

I've done it for all languages, but it's mostly problem for those
without much of a font difference, you can have no idea that the person
speaking to you is speaking in a different language than common. Plus
the icons are cool, and barely take up any screen space.

![image](https://github.com/user-attachments/assets/922e4695-b7f2-48b7-bbd7-f4cacf182dfb)

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->