Migrant Roles III (#2199)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds the following migrant roles:

- The Candy Man
Baothan. Minimal combat skills. Spawns with 24 instances of narcotics to
peddle to the townspeople. Looks like a thief to blend in, but has a
special title (Candyman/woman). Baotha's strongest crackhead.

![image](https://github.com/user-attachments/assets/a2b60bd5-e578-4e09-a0ba-8579acd5f21c)



- Ratcatcher
Horribly geared peasant. Average in polearms, skilled in shields to
compensate for his lack of defensive traits. Kills rats. Works for the
crown. Wants to die. Will probably die.

![image](https://github.com/user-attachments/assets/9022c1d6-3334-44e0-8313-6f2c5bfccce2)


- Urchin Wave
A gang of orphans (3-1). Weak/Average combat skills. Give the Matron the
biggest migrane as you get your brains splattered on the pavement.
Leader gets a knife, normal ones get a cudgel. Normal ones are also more
on the weak side. The leader gets a fancy hat to signify he's the boss.
Assume nobody would hurt an orphan as you mug a warrior and get
eviscerated.

![image](https://github.com/user-attachments/assets/d90e928d-26d2-4f8a-a330-0b897a8f8885)


## Why It's Good For The Game

A-lot of migrant roles right now require a large amount of people, or
are just pilgrim roles. These give more gimmicks to folks who might just
want a direction to go in, rather than having to play something freeform
like adventurer.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->