The Kaledonian Content Pack Pt. 1 (#2146)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This PR adds a number of items and subclasses associated with the nation
of Kaledon.


https://docs.google.com/document/d/1c0EI-lhsWS0uCeV_-RdjgOaW0uyHzmRu3J0flqWApac/edit?tab=t.0

It consists of

Sea/Coastal Elf Vikingrs - Adventurer
Gallowglass - Mercenary Claymore Wielder
Kern - Basic Mercanary (Polearm/Bow)

It also adds 3 types of Claymore in steel, iron and decorated, the
decorated is presently unused.

It also adds a be special trait for a meme greatsword, in the form of
the Sword from Berserk's Golden Age arc, it is heavily balanced around
being insanely powerful, but extremely stamina draining. The only way to
obtain this at present, is via being a beggar and rolling a rare special
trait.

Also updates the Merc text to mention the gaffer, not merchant as the
person to go to for jobs.



<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

Improves upon the game’s lore by introducing a variety of items and
subclasses from a region.
Adds more depth to adventurers and mercenaries by adding themed
subclasses
Gives elves a big W in the form of a combat-focused class


https://docs.google.com/document/d/1c0EI-lhsWS0uCeV_-RdjgOaW0uyHzmRu3J0flqWApac/edit?usp=sharing
- Kaledon's lore.


![dreamseeker_o1eLlNXqHc](https://github.com/user-attachments/assets/24124cec-8d6e-4364-b8c3-d498c16a692e)

![dreamseeker_zE60bEiMh3](https://github.com/user-attachments/assets/50b09df7-e535-412b-a6bb-dfdc2782ff9a)

![dreamseeker_e2pFCGwyr8](https://github.com/user-attachments/assets/bf25dfd9-660d-4ccf-b363-e3c9173d4421)

![dreamseeker_tT69QaC607](https://github.com/user-attachments/assets/ccf242ed-8a22-41b3-9fa6-37c7c4743aa9)

![dreammaker_Z3VYB16SY8](https://github.com/user-attachments/assets/f90004f5-3f49-42c6-93fd-db0b38446a5b)


<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>